remove: lancherDialog class from dashpva.py add: launcherDialog to viewer as it's own file in viewer folder
